First attempt to actually generate a comparison. The controlling syntax looks exactly like 'expView' commands, with the exception that we use 'expCompare' as the command name.
There are semantic differences between the two commands.
ExpView combines all the data samples selected in <target_list>,
where expCompare treats each specification as a separate set of
data items and generates the compare by placing each set in
separate columns of the final view.

The column headers are expanded by pre-pending the original
headers with the selection specification information.

Example:  (pcsamp experiment)
  expView -h mcr448, mcr449

combines all data into 1 column and creates the header:
           CPU Time  Function Name

  expCompare -h mcr448, mcr449

separates the data into 2 columns and creates the header:
 -h mcr448: CPU Time  -h mcr449: CPU Time  Function Name

The drawback of this approach is that the headers can become
too long to fit into the allowed width, and will be truncated.
We may wish to go to a multi-line header, in the future.


Also, we will, eventually, want to support more general syntax but will
need to fully define <Compare_Sets> before we have a handle in it.