Bugzilla #469
This change will hopefully centralize the search path algorithm
for OpenSpeedShop. The file libopenss-framework/OpenSSPath.cxx
is actually included and the function within it is a static.
This is because this algoritm needs to be accessed individually
from openss, the cli and the framework depending on how lowlevel
the toolmaker want to go.

For example, the python scripting API uses the cli, but not
openss. I believe some of the framework tests skip the cli.

Because the file is included and not compiled separately, I
am not sure where it needs to get registered so it will be
added in the source distribution.

Jack