(Fix per gui feedback...)
To navigate to a new function and gets its butterfly view we
needed to select and right click to get the metric menu and ask for
butterfly view.

Since we are already in a butterfly view, could we just use the arrow down or
up key (which is mapped and seems to highlight one of the entries) and
then hit return to get the new butterfly view for that position.

also, if we up/down arrow past eaither end, could the current butterfly
view update itself. right now we get at most 3 entries.

the gist of this email is that we need to use a only gui gestures (ie. popup
menu) to get the butterfly for a selected function. It's a bit cumbersome (and
the remote X window makes it even more so).





The StatsPanel, now accepts 'returnPress' events and attempts to handle them meaninfully based on the 'context' of the view.