This fixes cases for options that don't have to have an argument like expClose where the only argument is an error:
    expclose -x fkjakfja

This was caught as a parsing error, but not identified
with any specific command, thus you didn't get the
correct grammar for the command listed in the message.

This change does cause yacc to complain about reduce conflicts,
but they are harmless because either route it takes works fine.

Jack