Add initial code for IOT view.
At the moment, it is just a place holder waiting for the IOT collector to
show up.  Once that happens, the only major change required is to go into
the views/iot_view.cxx module and change IODetail to IOTDetail.  There's
even a chance that this is the only required change to ge the basic info
to come out.