This puts in the command "cViewCluster" for testing.
I also put in a check at CLI startup to make sure the
command enum and the command array are in sync along
with instructions on what files need to be mucked with
to add a command.

I forgot to add instructions for adding the command
to python and I also haven't added any testing for
any of the CV commands.

I have misgivings about this particular command still
in terms of what we are trying to achieve with it. Can
it be described in terms of a series of other building
block commands? If not, why not. This command does not
seem like a building block command. If it is, please explain
why it is. If it isn't, what are the building block commands
that make it up.

Do we need to add some attributes to other
commands to say we are talking about clusters?

We do something similar for MPI, -v mpi.

What will a list of cv_ids do for our customers if there
are hundreds of nodes? Do we assume that the return values
will only be useful to script and gui users?

Me thinks that the concept of custom views needs a few
more rounds. It is as if the only the gui can package
and manipulate the view/dataset for it's real potential.

I add "datasets" because for the script users they can
get semi-raw data back defined by these "views".

The concept of gathering and defining a library of views/datasets
is very powerful, but only if one can visualize the library and
make sub groups of views/datasets.


I guess I still just don't get it.

Jack