This is a temporary change until I get home to my python books.
The debugging workaround I put in is to NOT preempt
showtraceback(self). I do want to preempt it, but I
also want to dump the traceback.

The second part is setting in ~/.qt/openspeedshoprc

    allowPythonCommands=true

It is off by default because, and this is a good
reason, because preparsing is mostly broken and
even if it wasn't or you disagree, putting the
innocent into a python environment with its indentation
sensitivity is not good business.

We may want to consider another way of invoking StartDPCL
if this is to become mainstream.

Jack

aptop [159] openss -cli
openss>>Welcome to OpenSpeedShop 0.875
openss>>StartDPCL("mpirun -np 4 *", "/usr/local")
Traceback (most recent call last):
  File "<console>", line 1, in ?
  File "SS_Preparse.py", line 855, in StartDPCL
    raise RuntimeError("Failed to locate the DPCL Daemon.")
RuntimeError: Failed to locate the DPCL Daemon.
openss>>exit