This is in response to a non-bugzilla request for experiment interval view support. Current the interval list is just used for the view commands, but could easily be used for other commands as well.
I have NOT made any changes to the runtime help
for this. That will need to be done.

I have NOT added this to any non-view commands yet.
Please tell me you want it added to which commands.

<interval_spec>  ::= "-I" [<attribute>] <interval_list>
<attribute>      ::= "%"|| "s" || "ms" || "us" || "ns"
<interval_list>  ::= <interval_range>[,<interval_range>]*
<interval_range> ::= <interval_val> ":" <interval_val>
<interval_val>   ::= <int> || <float>

My guess is that <int> should become long long here.

Let me know if I messed up the definition and or the
code in general.

Jack