My earlier change to allow the gentle user move the openss directories to a non-standard location cause gdb to have fits. This commit will bypass that change with the setting of an environment variable:
    OPENSS_DEBUG_OPENSS

It doesn't matter what you set it to.

I picked this name after recursively grepping all instances
of getenv() in openspeedshop. If you do not approve of the
choice please bring up your objections.

Where do we document the various environment variables we
use? This one will definately need to be documented for the
OpenSpeedShop developer.

Regards,

Jack