Quoting a string was not working for some reason. I added another rule to lex.
The problem we have in the parser is that our syntax is rather
complicated and possibly context sensitive while our parser
rules are trying to be as context insensitive as possible.

Sometimes we want names to seperated by things like double colons
and other times there are names that have double colons as part
of the names such as:

    ggFineSpectrum::SetRange

For now we require these names to be surrounded by quotes
until we understand our own rules which continue to evolve.

Jack