This is to use lib64 instead of lib when appropriate. This is currently decided in configure.ac for x86-64 vs everything else.
This will currently affect only the openss build.
For the framework or cli standalone builds the define
for LIB_DIR_NAME will have to be put in their respective
Makefile.am files. I haven't done this yet because I want
to test the change on an x86-64 system first.

It's possible that we will want to check both lib and lib64
for some systems.

Jack