"Update loop and vector instruction control
Update the loop and vector instruction code for better control and
messaging when turning off the detection of loops and vector instructions.
Multiple messages were output before.   Also, add an environment variable:
OPENSS_VINSTR_OPSIZE to control the size of the vector instruction operands to
look for.  The allowed options are 128, 256, and 512.  If 128 is chosen O|SS
will report vector instructions with operand sizes of 128 and above, if 256 is
chosen O|SS will report vector instructions with operand sizes of 256 and above."