Add new options to the O|SS driver convenience script
Turn off loop detection and vector instruction detection
by default and add options to the convenience scripts that
allows users to turn on loop detection and vector instruction
detection when desired.

The new options are as follows:
# Enable the executable post-processing for loop detection
      --loops)
# Disable the executable post-processing for loop detection
       --noloops)
# Enable the executable post-processing for vector instruction
# detection for instructions with 128 bit or larger operands
       --vinstr128)
# Enable the executable post-processing for vector instruction
# detection for instructions with 256 bit or larger operands
       --vinstr256)
# Enable the executable post-processing for vector instruction
# detection for instructions with 512 bit or larger operands
       --vinstr512)

Also add help documentation in the convenience scripts
so users are aware of the options.  Provide messages when
the scripts are running, so users are are that detection
is on or off.